implement exclusive virtualhost
SOGo is by default exposed to all (virtual)hosts, ie all FQDN/SOGo known
by apache.
If a FQDN is set in the sogod virtualhost porp  SOGo is expossed
exclusive through this virtualhost and is accessible for other
(virtual)hosts.
If present, in this order: the Certificate set in the Certificate prop
-or Sogo.crt -or localhost.crt is used